Rename namespace PG.Logic.Passwords.Loader > PG.Logic.Passwords.Loaders
