+++ pg-logic/Passwords/Generators/PasswordGeneratorFactory.cs
@@ -2,7 +2,7 @@
 using PG.Entities.WordTrees;
 using PG.Logic.Common;
 using PG.Logic.Passwords.Generators.Entities;
-using PG.Logic.Passwords.Loader;
+using PG.Logic.Passwords.Loaders;
 using PG.Shared.Services;
 
 namespace PG.Logic.Passwords.Generators
