+++ pg-logic/Passwords/Loaders/DictionaryLoaderFactory.cs
@@ -4,7 +4,7 @@ using PG.Data.Files.DataFiles.WordTrees;
 using PG.Entities.Files;
 using System.Text;
 
-namespace PG.Logic.Passwords.Loader
+namespace PG.Logic.Passwords.Loaders
 {
 	public class DictionaryLoaderFactory(IServiceProvider provider) : IDictionaryLoaderFactory
 	{
