+++ pg-logic/Passwords/Loaders/IDictionaryLoaderFactory.cs
@@ -2,7 +2,7 @@
 using PG.Entities.Files;
 using System.Text;
 
-namespace PG.Logic.Passwords.Loader
+namespace PG.Logic.Passwords.Loaders
 {
 	public interface IDictionaryLoaderFactory
 	{
