+++ pg-tests/Interface/PasswordGeneration/PassGenieParserTests.cs
@@ -2,7 +2,7 @@
 using PG.Data.Files.DataFiles;
 using PG.Interface.Command.PasswordGeneration;
 using PG.Logic.Passwords.Generators;
-using PG.Logic.Passwords.Loader;
+using PG.Logic.Passwords.Loaders;
 using PG.Shared.Services;
 
 namespace PG.Tests.Interface.PasswordGeneration
